stream/requirements.txt
redis==7.3.0
python-dotenv==1.1.1
pydantic==2.11.9
python-dateutil==2.9.0.post0
alpaca_trade_api==3.2.0
